+++ BoardBackend/src/main/java/org/example/boardbackend/model/dto/board/club/ClobUtils.java
@@ -1,0 +1,32 @@
+package org.example.boardbackend.model.dto.board.club;
+
+import java.io.IOException;
+import java.io.Reader;
+import java.sql.Clob;
+import java.sql.SQLException;
+
+/**
+ * packageName : org.example.boardbackend.model.dto.board.club
+ * fileName : ClobUtils
+ * author : BALLBAT
+ * date : 2024-06-04
+ * description :
+ * 요약 :
+ * <p>
+ * ===========================================================
+ * DATE            AUTHOR             NOTE
+ * -----------------------------------------------------------
+ * 2024-06-04         BALLBAT          최초 생성
+ */
+public class ClobUtils {
+    public static String clobToString(Clob clob) throws SQLException, IOException {
+        StringBuilder sb = new StringBuilder();
+        try (Reader reader = clob.getCharacterStream()) {
+            int c;
+            while ((c = reader.read()) != -1) {
+                sb.append((char) c);
+            }
+        }
+        return sb.toString();
+    }
+}
