+++ boardfrontend/public/index.html
@@ -21,6 +21,13 @@
       integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65"
       crossorigin="anonymous"
     />
+    <!-- 매칭 게시판 아이콘 -->
+    <link
+    rel="stylesheet"
+    href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css"
+    integrity="sha384-k6RqeWeci5ZR/Lv4MR0sA0FfDOMVlyL4N6Ry8FA1Fj5Nlb/ENt5YML4/iwdb6r5k"
+    crossorigin="anonymous"
+  />
     <!-- 일기예보 api -->
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" rel="stylesheet">
     <!-- 네이버 로그인 api -->
