+++ boardfrontend/src/views/club/clubBoardDetail.vue
@@ -1,76 +1,66 @@
 <template>
-  <div class="match-form-wrapper">
-    <h3 class="form-title">매칭 글 작성</h3>
-    <div class="form-container">
-      <div class="match-form">
-<input type="file" multiple @change="handlePhotoUpload" class="file-input" />
-        <div class="match-points">
-          <!-- 성별 선택 -->
-          <div class="select-wrapper">
-            <select v-model="matchDetails.sex" class="half-width-select">
-              <option disabled value="">성별 선택</option>
-              <option value="남자만">남자만</option>
-              <option value="여자만">여자만</option>
-              <option value="남녀 모두">남녀 모두</option>
-            </select>
-            <div class="selected-value">{{ matchDetails.sex }}</div>
+  <div style="background-color: #f2f2f2; height: 5000px">
+    <div class="container d-flex justify-content-center text-center">
+      <div v-if="clubBoard && !mapView" style="width: 80%; height: 500px; background-color: #4480fe">
+        <img :src="clubBoard.imgUrl" alt="club_photo" width="80%" height="500" class="d-inline-block align-text-top" />
+      </div>
+      <div id="map" style="width: 80%; height: 500px" ref="map" v-else></div>
+    </div>
+
+    <div class="container text-center mt-1">
+      <div class="container text-center">
+        <div class="lotto_new row row-cols-lg-2 gap-2 justify-content-center">
+          <div class="col match-point-container" style="background-color: #ffffff; height: 400px; max-width: 40%">
+            <p class="container text-left match-point-title">
+              매치 포인트
+            </p>
+            <div v-if="clubBoard" class="match-point-content">
+              <div class="match-point-item">
+                <i class="fas fa-venus-mars"></i>
+                <span>{{ clubBoard.sex }}</span>
+              </div>
+              <div class="match-point-item">
+                <i class="fas fa-futbol"></i>
+                <span>{{ clubBoard.peoplesMatch }} {{ clubBoard.matchForm }}</span>
+              </div>
+              <div class="match-point-item">
+                <i class="fas fa-users"></i>
+                <span>{{ clubBoard.minQuota }} ~ {{ clubBoard.maxQuota }}명</span>
+              </div>
+              <div class="match-point-item">
+                <i class="fas fa-running"></i>
+                <span>{{ clubBoard.material }}</span>
+              </div>
+            </div>
           </div>
-          <!-- 경기 인원 수 선택 -->
-          <div class="select-wrapper">
-            <select v-model="matchDetails.peoplesMatch" class="half-width-select">
-              <option disabled value="">경기 인원 수 선택</option>
-              <option value="5vs5">5vs5</option>
-              <option value="6vs6">6vs6</option>
-            </select>
-            <div class="selected-value">{{ matchDetails.peoplesMatch }}</div>
-          </div>
-          <!-- 경기 방식 선택 -->
-          <div class="select-wrapper">
-            <select v-model="matchDetails.matchForm" class="half-width-select">
-              <option disabled value="">경기 방식 선택</option>
-              <option value="3파전">3파전</option>
-              <option value="토너먼트">토너먼트</option>
-            </select>
-            <div class="selected-value">{{ matchDetails.matchForm }}</div>
-          </div>
-          <!-- 모집 인원 수 -->
-          <input type="text" v-model="matchDetails.minQuota" placeholder="최소 모집 인원 수" class="input-field" />
-          <input type="text" v-model="matchDetails.maxQuota" placeholder="최대 모집 인원 수" class="input-field" />
-          <!-- 준비물 선택 -->
-          <div class="select-wrapper">
-            <select v-model="matchDetails.material" class="half-width-select">
-              <option disabled value="">준비물 선택</option>
-              <option value="풋살화만">풋살화만</option>
-              <option value="운동화만">운동화만</option>
-              <option value="축구화만">축구화만</option>
-              <option value="풋살화/운동화">풋살화/운동화</option>
-              <option value="풋살화/축구화">풋살화/축구화</option>
-            </select>
-            <div class="selected-value">{{ matchDetails.material }}</div>
+          <div class="col" style="background-color: #ffffff; height: 400px; max-width: 40%">
+            news 2
+            <button @click="popUpMap">장소 사진 보기</button>
           </div>
         </div>
-        <!-- 시간 설정 -->
-        <div class="time-wrapper">
-          <div class="time-setting">
-            <label for="startTime" class="time-label">시작 시간 설정</label>
-            <input type="datetime-local" id="startTime" v-model="matchDetails.startTime" class="time-input" />
-          </div>
-          <div class="time-setting">
-            <label for="endTime" class="time-label">종료 시간 설정</label>
-            <input type="datetime-local" id="endTime" v-model="matchDetails.endTime" class="time-input" />
-          </div>
-          <div class="time-setting">
-            <label for="recruitmentDeadline" class="time-label">모집 마감 시간 설정</label>
-            <input type="datetime-local" id="recruitmentDeadline" v-model="matchDetails.recruitmentDeadline" class="time-input" />
+
+        <div class="d-flex justify-content-center">
+          <div class="row mt-3" style="background-color: #ffffff; width: 80%; height: 700px">
+            <p class="container text-left" style="text-align: left; margin: 20px 0 0 10px; font-size: 20px; letter-spacing: -1px; font-weight: 600;">
+              매치 데이터
+            </p>
+            <p v-if="clubBoard">{{ clubBoard.location }}</p>
           </div>
         </div>
-        <!-- 장소 및 기타 정보 입력 -->
-        <input type="text" v-model="matchDetails.location" placeholder="장소명 입력" class="input-field full-width" />
-        <input type="text" v-model="matchDetails.address" placeholder="주소 검색" readonly @click="openAddressSearch" class="input-field full-width" />
-        <input type="text" v-model="matchDetails.participationFee" placeholder="참가비 설정" class="input-field full-width" />
-        <textarea v-model="matchDetails.title" placeholder="구장 정보 입력" class="textarea-field"></textarea>
-        <textarea v-model="matchDetails.content" placeholder="매치 진행 방식 입력" class="textarea-field"></textarea>
-        <button @click="submitForm" class="submit-button">제출하기</button>
+
+        <div class="d-flex justify-content-center">
+          <div class="row mt-3" style="background-color: #ffffff; width: 80%; height: 700px; text-align: left">
+            <p class="container text-left" style="text-align: left; margin: 20px 0 0 10px; font-size: 20px; letter-spacing: -1px; font-weight: 600;">
+              매치 진행방식
+            </p>
+            <ul style="padding: 0 0 0 30px; font-size: 20px; padding-left: 50px">
+              <b>매치 규칙</b>
+              <li>모든 파울은 사이드라인에서 킥인</li>
+              <li>골키퍼에게 백패스 가능 손으로는 잡으면 안 돼요</li>
+              <li>사람을 향한 태클 금지</li>
+            </ul>
+          </div>
+        </div>
       </div>
     </div>
   </div>
@@ -78,222 +68,113 @@
 
 <script>
 import ClubBoardService from "@/services/board/club/ClubBoardService";
-import store from "@/store";
 
 export default {
   data() {
     return {
-      matchDetails: {
-        content: "",
-        location: "",
-        address: "",
-        participationFee: undefined,
-        startTime: "",
-        endTime: "",
-        recruitmentDeadline: "",
-        maxQuota: "",
-        minQuota: "",
-        peoplesMatch: "",
-        material: "",
-        matchForm: "",
-        title: "",
-        sex: "",
+      clubBoard: null,
+      options: {
+        center: { lat: 33.450701, lng: 126.570667 },
+        level: 3,
       },
-      imgFiles: [], // 여러 파일을 저장하기 위한 배열
+      address: "",
+      mapView: true,
     };
   },
   methods: {
-    handlePhotoUpload(event) {
-      this.imgFiles = Array.from(event.target.files);
-      console.log("Selected files:", this.imgFiles);
-       // 파일 크기 확인 로직 추가
-      const maxSize = 10 * 1024 * 1024; // 10MB
-      for (let file of this.imgFiles) {
-        if (file.size > maxSize) {
-          alert(`파일 크기가 너무 큽니다: ${file.name}`);
-          return;
-        }
-      }
-    },
-    openAddressSearch() {
-      if (typeof daum === "undefined") {
-        alert("Daum 우편번호 서비스를 로드할 수 없습니다.");
-        return;
-      }
-      new daum.Postcode({
-        oncomplete: (data) => {
-          this.matchDetails.address = data.address;
-        },
-      }).open();
-    },
-    async submitForm() {
+    async fetchClubBoardDetails() {
+      const clubBoardId = this.$route.params.clubBoardId;
       try {
-        let data = {
-          ...this.matchDetails,
-          userId: store.state.user.userId,
-        };
-
-        let formData = new FormData();
-        formData.append("data", JSON.stringify(data));
-        if (this.imgFiles.length > 0) {
-          this.imgFiles.forEach((file) => {
-            formData.append(`imgFiles`, file);
-          });
-        }
-
-        for (var pair of formData.entries()) {
-          console.log(pair[0] + ', ' + pair[1]);
-        }
-
-        let response = await ClubBoardService.createClub(formData);
+        console.log("Fetching details for clubBoardId:", clubBoardId);
+        const response = await ClubBoardService.getClubOnce(clubBoardId);
         console.log("Response data:", response.data);
-        this.$router.push("/club/club-board");
-      } catch (e) {
-        console.error("Error:", e);
-        if (e.response) {
-          console.error("Response data:", e.response.data);
-          console.error("Response status:", e.response.status);
-          console.error("Response headers:", e.response.headers);
-        } else if (e.request) {
-          console.error("Request data:", e.request);
-        } else {
-          console.error("Error message:", e.message);
-        }
-        console.error("Config:", e.config);
+        this.clubBoard = response.data[0]; // 첫 번째 요소를 clubBoard에 할당
+        this.address = this.clubBoard.address;
+        this.retrieveMap(this.address);
+      } catch (error) {
+        console.error("Error fetching club board details:", error);
       }
     },
+    popUpMap() {
+      this.mapView = !this.mapView;
+      if (this.mapView) {
+        this.$nextTick(() => {
+          setTimeout(() => {
+            this.retrieveMap(this.address);
+          }, 0);
+        });
+      }
+    },
+    retrieveMap(address) {
+      let kakao = window.kakao;
+      var container = this.$refs.map;
+      const { center, level } = this.options;
+
+      var map = new kakao.maps.Map(container, {
+        center: new kakao.maps.LatLng(center.lat, center.lng),
+        level,
+      });
+
+      var geocoder = new kakao.maps.services.Geocoder();
+
+      geocoder.addressSearch(address, function (result, status) {
+        if (status === kakao.maps.services.Status.OK) {
+          var coords = new kakao.maps.LatLng(result[0].y, result[0].x);
+
+          var marker = new kakao.maps.Marker({
+            map: map,
+            position: coords,
+          });
+
+          var infowindow = new kakao.maps.InfoWindow({
+            content: '<div style="width:150px;text-align:center;padding:6px 0;">경기장소</div>',
+          });
+          infowindow.open(map, marker);
+
+          map.setCenter(coords);
+        }
+      });
+    }
   },
-  mounted() {
-    console.log("Store state user:", store.state.user.userId);
-  },
+  async mounted() {
+    await this.fetchClubBoardDetails();
+  }
 };
 </script>
 
 <style>
-.match-form-wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
+.match-point-container {
   padding: 20px;
-  background-color: #f5f5f5;
 }
 
-.form-title {
-  margin-bottom: 20px;
-  letter-spacing: 1.5px;
-  color: #999999;
-  font-weight: bold;
-}
-
-.form-container {
-  width: 100%;
-  max-width: 1000px;
-  background-color: #ffffff;
-  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.1);
-  padding: 40px;
-  border-radius: 10px;
-}
-
-.match-form {
-  display: flex;
-  flex-direction: column;
-}
-
-.file-input {
-  color: #999999;
-  margin-bottom: 20px;
-}
-
-.select-wrapper {
-  display: flex;
-  align-items: center;
-  margin: 10px 0;
-}
-
-.half-width-select {
-  flex: 1;
-  padding: 10px;
-  border: 1px solid #cccccc;
-  border-radius: 4px;
-}
-
-.selected-value {
-  margin-left: 10px;
-  padding: 10px;
-  border: 1px solid #162b59;
-  border-radius: 4px;
-  color: #162b59;
-  font-weight: bold;
-  text-align: center;
-  width: 150px;
-}
-
-.input-field {
-  height: 50px;
-  padding: 0 20px;
-  border: 1px solid #cccccc;
-  border-radius: 4px;
-  margin-bottom: 10px;
-  box-sizing: border-box;
-}
-
-.input-field.full-width {
-  width: 100%;
-}
-
-.input-field::placeholder {
-  text-align: center;
-}
-
-.time-wrapper {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 20px;
-}
-
-.time-setting {
-  display: flex;
-  flex-direction: column;
-  width: 32%;
-}
-
-.time-label {
-  font-size: 14px;
-  color: #999999;
-  font-weight: bold;
-  margin-bottom: 5px;
-}
-
-.time-input {
-  height: 50px;
-  padding-left: 20px;
-  border: 1px solid #cccccc;
-  border-radius: 4px;
-}
-
-.textarea-field {
-  height: 100px;
-  padding: 10px 20px;
-  border: 1px solid #cccccc;
-  border-radius: 4px;
-  margin-bottom: 10px;
-  box-sizing: border-box;
-}
-
-.submit-button {
-  height: 50px;
-  font-weight: bold;
-  color: #ffffff;
+.match-point-title {
+  text-align: left;
+  margin: 20px 0 0 10px;
   font-size: 20px;
-  background-color: #162b59;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
+  letter-spacing: -1px;
+  font-weight: 600;
+}
+
+.match-point-content {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
   margin-top: 20px;
 }
 
-.submit-button:hover {
-  background-color: #0d1f3c;
+.match-point-item {
+  display: flex;
+  align-items: center;
+  margin-bottom: 10px;
+}
+
+.match-point-item i {
+  margin-right: 10px;
+  width: 30px;
+  height: 30px;
+  font-size: 30px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 </style>
